requirements.txt
streamlit
requests
beautifulsoup4
gensim
numpy
groq
python-dotenv
fpdf
faiss-cpu
langchain-community
